Move DocumentConverter to own folder
